integrate upload into github tab
