+++ components/bpmn-q/test/tests/editor/plugin.spec.js
@@ -54,7 +54,7 @@ describe('Test plugins', function () {
                 expect(extensions['planqk']).to.not.be.undefined;
                 expect(transfButtons.length).to.equal(3);
                 expect(buttons.length).to.equal(2);
-                expect(tabs.length).to.equal(5);
+                expect(tabs.length).to.equal(4);
                 expect(styles.length).to.equal(3);
             });
         });
